code review: remove storage interface and storage service
