+++ config/default/SyncPackageService.conf.php
@@ -3,6 +3,4 @@
  * Default config header created during install
  */
 
-return new oat\taoSync\package\SyncPackageService(array(
-    'storage' => new oat\taoSync\package\storage\SyncFileSystem()
-));
+return new oat\taoSync\package\SyncPackageService();
